fix(CoreDataManager) : 현재 위치 정보는 삭제되지 않도록 수정

+++ Yellow/CoreData/CoreDataManager+WeatherLocation.swift
@@ -129,6 +129,8 @@ extension CoreDataManager where ManagedObject == WeatherLocation {
             return
         }
         let weatherLocation = fetchedResultsController.object(at: IndexPath(row: index, section: 0))
+        guard !weatherLocation.isCurrentLocation else { return }
+        
         context.delete(weatherLocation)
         fetchedObjects.remove(at: index)
